% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MismatchFinder.R
\name{MismatchFinder}
\alias{MismatchFinder}
\title{ScisorWiz Mismatch Finder Function}
\usage{
MismatchFinder(BAM, fasta, gencodeAnno, gene, outputDir)
}
\arguments{
\item{BAM}{A sorted .bam file for the dataset of interest}

\item{fasta}{A reference genome .fasta file for cross referencing data in
.bam file to reference genome}

\item{gencodeAnno}{Gencode annotation file from which data files are created}

\item{gene}{User-specified gene of interest}

\item{outputDir}{User-specified directory to store output from the function}
}
\value{
Mismatch file with positions of SNPs, insertions, and deletions
organized by read
}
\description{
Cross references data from .bam file against a reference genome
to determine mismatches in the genetic code. These mismatches include
Single Nucleotide Polymorphisms (SNPs), nucleotide insertions, and
nucleotide deletions. These mismatches are listed per read to be later
included in the ScisorWiz Pipeline function.
}
