% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScisorWiz_2File.R
\name{ScisorWiz_2File}
\alias{ScisorWiz_2File}
\alias{ScisorWiz2File}
\title{ScisorWiz 2 File pipeline}
\usage{
ScisorWiz_2File(gencodeAnno, gffInput, genesInput, cellTypeFile, gene,
ci, mismatchCutoff, outputDir, mismatchFile, interactive)
}
\arguments{
\item{gencodeAnno}{Gencode annotation file from which data files are created}

\item{gffInput}{Gff file housing read data: chromosome, genome, readtype,
start, end, strand direction, and readID}

\item{genesInput}{Genes file housing read data: readID, geneID, read and gene
quality}

\item{cellTypeFile}{Tab-separated file housing user specified cell types and
desired color of reads of those cell types for the output plot}

\item{gene}{User-specified gene of interest}

\item{ci}{User-specified confidence interval for reads to be considered
alternate exons. Default is .05.}

\item{mismatchCutoff}{User-specified cutoff for SNV inclusion rate. Default
is .05.}

\item{outputDir}{User-specified directory to store all output from the
pipeline}

\item{mismatchFile}{Output of MismatchFinder function if used. Default is
NULL.}

\item{interactive}{Yes (y) or No (n) for creating an interactive html plot
file using a jpg produced from user input. Default is No (n).}
}
\value{
Plot visualizing isoform expression of a gene of interest among any
number of user-specified cell types
}
\description{
This is the script to run the entire pipeline if the user inputs
a gff and genes.gz file as their data. Uses the data from single-cell long-
read RNA sequencing to visualize the isoform expression patterns of a single
gene across any number of celltypes. Data is clustered by intron chain. The
user has option to include a mismatch file by first running necessary files
through the MismatchFinder function. User can also choose to create an
interactive plot for exploratory purposes by setting the "interactive"
argument equal to "y".
}
