% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScisorWiz_AllInfo.R
\name{ScisorWiz_AllInfo}
\alias{ScisorWiz_AllInfo}
\alias{ScisorWizAllInfo}
\title{ScisorWiz AllInfo File pipeline}
\usage{
ScisorWiz_AllInfo(gencodeAnno, AllInfoInput, cellTypeFile, gene,
cluster, ci, mismatchCutoff, outputDir, mismatchFile, interactive)
}
\arguments{
\item{gencodeAnno}{Gencode annotation file from which data files are created}

\item{AllInfoInput}{File generated from the scisorseqr package that houses
all data in a uniquely organized fashion: readID, geneID, celltype, barcode,
intron chain, etc.}

\item{cellTypeFile}{Tab-separated file housing user specified cell types and
desired color of reads of those cell types for the output plot}

\item{gene}{User-specified gene of interest}

\item{cluster}{User-specified clustering method (1 = intron chain, 2 =
TSS site, 3 = PolyA site, 4 = intron chain, TSS site, and PolyA site).
Default is 1 (intron chain)}

\item{ci}{User-specified confidence interval for reads to be considered
alternate exons. Default is .05.}

\item{mismatchCutoff}{User-specified cutoff for SNV inclusion rate. Default
is .05}

\item{outputDir}{User-specified directory to store all output from the
pipeline}

\item{mismatchFile}{Output of MismatchFinder function if used. Default is
NULL.}

\item{interactive}{Yes (y) or No (n) for creating an interactive html plot
file using a jpg produced from user input. Default is No (n).}
}
\value{
Plot visualizing isoform expression of a gene of interest among any
number of user-specified cell types
}
\description{
This is the script to run the entire pipeline if the user inputs
an AllInfo File as their data. AllInfo file is an output of the scisorseqr
pipeline. Uses data from single-cell long-read RNA sequencing to visualize
the isoform expression patterns of a single gene across any number of cell
types. Data is clustered in one of 4 ways: by intron chain of the reads, by
TSS site, by PolyA site, or by all three clustering methods. User has option
to include a mismatch file by first running necessary files through the
MismatchFinder function. User can choose to create an interactive plot for
exploratory purposes by setting the "interactive" argument equal to "y".
}
