% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScisorWiz_2File.R
\name{ScisorWiz_2File}
\alias{ScisorWiz_2File}
\alias{ScisorWiz2File}
\title{ScisorWiz 2 File pipeline}
\usage{
ScisorWiz_2File(gencodeAnno, gffInput, genesInput, cellTypeFile, gene,
ci, mismatchCutoff, outputDir, mismatchFile)
}
\arguments{
\item{gencodeAnno}{Gencode annotation file from which data files are created}

\item{gffInput}{Gff file housing read data: chromosome, genome, readtype,
start, end, strand direction, and readID}

\item{genesInput}{Genes file housing read data: readID, geneID, read and gene
quality}

\item{cellTypeFile}{File housing user specified celltypes and desired color
of reads of those celltypes for the output plot}

\item{gene}{User-specified gene of interest}

\item{ci}{User-specified confidence interval for reads to be considered
alternate exons. Default is .05.}

\item{mismatchCutoff}{User-specified cutoff for SNV inclusion rate. Default
is .05.}

\item{outputDir}{User-specified directory to store all output from the
pipeline}

\item{mismatchFile}{Output of MismatchFinder function if used. Default is
NULL.}
}
\value{
Plot visualizing isoform expression of the gene of interest among up
to 6 user-specified cell types
}
\description{
This is the script to run the entire pipeline if the user enters
a gff and genes.gz file as their data. Uses the data from single-cell long-
read RNA sequencing to visualize the isoform expression patterns of a single
gene across up to six celltypes. The data is clustered by intron chain. User
has option to include a mismatch file by first running necessary files
through the MismatchFinder function.
}
