% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScisorWiz_AllInfo.R
\name{ScisorWiz_AllInfo}
\alias{ScisorWiz_AllInfo}
\alias{ScisorWizAllInfo}
\title{ScisorWiz AllInfo File pipeline}
\usage{
ScisorWiz_AllInfo(gencodeAnno, AllInfoInput, cellTypeFile, gene,
cluster, ci, mismatchCutoff, outputDir, mismatchFile, zoom, interactive)
}
\arguments{
\item{gencodeAnno}{Gencode annotation file from which data files are created}

\item{AllInfoInput}{File generated from the scisorseqr package that houses
all data in a uniquely organized fashion: readID, geneID, celltype, barcode,
intron chain, etc.}

\item{cellTypeFile}{File housing user specified celltypes and desired color
of reads of those celltypes for the output plot}

\item{gene}{User-specified gene of interest}

\item{cluster}{User-specified clustering method (1 = intron chain, 2 =
TSS site, 3 = PolyA site, 4 = intron chain, TSS site, and PolyA site).
Default is 1 (intron chain)}

\item{ci}{User-specified confidence interval for reads to be considered
alternate exons. Default is .05.}

\item{mismatchCutoff}{User-specified cutoff for SNV inclusion rate. Default
is .05}

\item{outputDir}{User-specified directory to store all output from the
pipeline}

\item{mismatchFile}{Output of MismatchFinder function if used. Default is
NULL.}

\item{zoom}{Yes (y) or No (n) for zooming into a user-specified window on the
plot. Default is No (n).}

\item{interactive}{Yes (y) or No (n) for creating an interactive plot session
using the pdf produced from user input. Default is No (n).}
}
\value{
Plot visualizing isoform expression of the gene of interest among up
to 6 user-specified cell types
}
\description{
This is the script to run the entire pipeline if the user enters
an AllInfo File as their data. AllInfo file is an output of the scisorseqr
pipeline. Uses the data from si  ngle-cell long-read RNA sequencing to
visualize the isoform expression patterns of a single gene across up to six
celltypes. The data is clustered in one of 4 ways: by intron chain of the
reads, by TSS site, by PolyA site, or by PolyA and TSS sites. User has option
to include a mismatch file by first running necessary files through the
MismatchFinder function.
}
